% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sophisthse.R
\name{decyrillic}
\alias{decyrillic}
\title{Replace cyrillic letters by corresponding latin letters}
\usage{
decyrillic(x)
}
\arguments{
\item{x}{the character vector}
}
\value{
clean character vector
}
\description{
Replace cyrillic letters by corresponding latin letters
}
\details{
Replace cyrillic letters by corresponding latin letters
}
\examples{
sophisthse:::decyrillic(intToUtf8(1057))
}

