% Generated by roxygen2 (4.0.2): do not edit by hand
\name{remove_slash_junk}
\alias{remove_slash_junk}
\title{Remove slash junk}
\usage{
remove_slash_junk(x)
}
\arguments{
\item{x}{the character vector}
}
\value{
clean character vector
}
\description{
Remove slash junk
}
\details{
Remove slash junk
}
\examples{
remove_slash_junk("xxx \\n yyy")
}

