% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sophisthse.R
\name{rus2num}
\alias{rus2num}
\title{Convert string with a number in Russian tradition in numeric}
\usage{
rus2num(x)
}
\arguments{
\item{x}{the string with the number}
}
\value{
numeric the number converted from the string
}
\description{
This function is useful for automatic conversion of strings to numeric
}
\details{
Russian standards prescribes to use comma as a decimal separator.
This function removes spaces and converts string to number.
}
\examples{
rus2num('34 345,34')
}

