% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{series_info}
\alias{series_info}
\title{Description of time series available at sophist.hse.ru}
\format{
data.frame with 384 rows and 8 columns:
\describe{
  \item{table}{name of table that contains time series}
  \item{tsname}{short time series name}
  \item{freq}{frequency, 1 for yearly, 4 for quarterly, 12 for monthly data}
  \item{unit}{measurement unit}
  \item{fullname}{full time series name}
  \item{methodology}{methodology of calculation}
  \item{source}{source of time series}
  \item{comments}{further comments for time series}
}
}
\source{
downloaded from \url{http://sophist.hse.ru/hse/nindex.shtml} on 15.06.2016
}
\usage{
series_info
}
\description{
A dataset containing description of time series available at \url{http://sophist.hse.ru/hse/nindex.shtml}.
Downloaded on 15.06.2016.
}
\keyword{datasets}
