% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sophisthse.R
\name{set_variable_labels}
\alias{set_variable_labels}
\title{Set variable labels of a data.frame}
\usage{
set_variable_labels(df, labels)
}
\arguments{
\item{df}{data.frame}

\item{labels}{character vector of variable names}
}
\value{
data.frame with labelled variables
}
\description{
Set variable labels of a data.frame
}
\details{
Set variable labels of a data.frame. They will be nicely visible
with `View()` in Rstudio.
}
\examples{
cars2 <- set_variable_labels(cars,
  labels = c("Speed (mph)", "Stopping distance (ft)"))
}

