% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sophisthse.R
\docType{package}
\name{sophisthse}
\alias{sophisthse}
\alias{sophisthse-package}
\title{sophisthse}
\usage{
sophisthse(series.name = "IP_EA_Q", output = c("zoo", "data.frame"))
}
\arguments{
\item{series.name}{the names of the time series, i.e. 'WAG_Y'}

\item{output}{the desired output format, either 'zoo' or 'data.frame'}
}
\value{
data.frame with the corresponding time series
}
\description{
sophisthse

This function obtains multivariate time series from sophist.hse.ru
}
\details{
The output may be choosen to be 'zoo' or 'data.frame'. Metadata is saved
into the attribute 'metadata'.
}
\examples{
df <- sophisthse('IP_EA_Q')
df <- sophisthse('WAG_Y')
}
\author{
Boris Demeshev
}

