% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sophisthse.R
\name{sophisthse0}
\alias{sophisthse0}
\title{Obtain time series from sophist.hse.ru}
\usage{
sophisthse0(series.name = "IP_EA_Q", ...)
}
\arguments{
\item{series.name}{the names of the time series, i.e. 'WAG_Y'}

\item{...}{further arguments passed into getURL. One may use them to work with proxy.}
}
\value{
data.frame with the corresponding time series
}
\description{
This function obtains univariate time series from sophist.hse.ru
}
\details{
The output may be choosen to be 'ts', 'zoo' or 'data.frame'. Metadata is saved
into the attribute 'metadata'.
}
\examples{
df <- sophisthse0('IP_EA_Q')
df <- sophisthse0('WAG_Y')
}

