% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sophisthse.R
\name{sophisthse_metadata}
\alias{sophisthse_metadata}
\title{Get metadata from downloaded time series}
\usage{
sophisthse_metadata(df)
}
\arguments{
\item{df}{downloaded multivariate time series or data.frame}
}
\value{
data.frame with information
}
\description{
Get metadata from downloaded time series
}
\details{
Get metadata from downloaded time series
}
\examples{
df <- sophisthse0('WAG_Y')
sophisthse_metadata(df)
}
