% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sophisthse.R
\name{remove_slash_junk}
\alias{remove_slash_junk}
\title{Remove slash junk}
\usage{
remove_slash_junk(x)
}
\arguments{
\item{x}{the character vector}
}
\value{
clean character vector
}
\description{
Remove slash junk
}
\details{
Remove slash junk
}
\examples{
sophisthse:::remove_slash_junk('xxx \n yyy')
}
