% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sophisthse.R
\name{requested_freq}
\alias{requested_freq}
\title{Guess series frequency from it's name}
\usage{
requested_freq(series.name)
}
\arguments{
\item{series.name}{the name of a time series}
}
\value{
guessed frequency (1/4/12)
}
\description{
Guess series frequency from it's name for sophist.hse.ru
}
\details{
Guess series frequency from it's name
by checking presence of letters Y, M, Q
}
\examples{
requested_freq("WAG_Y")
}

