% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sophisthse.R
\name{sophisthse}
\alias{sophisthse}
\title{Obtain multivariate time series from sophist.hse.ru}
\usage{
sophisthse(
  series.name = "IP_EA_Q",
  output = c("ts", "zoo", "data.frame", "tsibble"),
  ...
)
}
\arguments{
\item{series.name}{the names of the time series, i.e. 'WAG_Y'}

\item{output}{the desired output format, either 'ts', 'zoo', 'data.frame' or 'tsibble'.}

\item{...}{further arguments passed into getURL. One may use them to work with proxy.}
}
\value{
data.frame with the corresponding time series
}
\description{
This function obtains multivariate time series from sophist.hse.ru
}
\details{
The output may be choosen to be 'ts', 'zoo' or 'data.frame'. Metadata is saved
into the attribute 'metadata'.
}
\examples{
df <- sophisthse('IP_EA_Q')
df <- sophisthse('WAG_Y')
}
