% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{sophisthse}
\alias{sophisthse}
\alias{sophisthse-package}
\title{sophisthse}
\usage{
sophisthse(series.name = "IP_EA_Q", output = c("zoo", "data.frame"))
}
\arguments{
\item{series.name}{the names of the time series, i.e. "WAG_Y"}

\item{output}{the desired output format, either 'zoo' or 'data.frame'}
}
\value{
data.frame with the corresponding time series
}
\description{
sophisthse

This function obtains multiple time series from sophist.hse.ru
}
\details{
If many time series are requested at once they should be of the same frequency.
}
\examples{
df <- sophisthse("IP_EA_Q")
df <- sophisthse("WAG_Y")
}
\author{
Boris Demeshev
}

