% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sophisthse.R
\name{sophisthse_tables}
\alias{sophisthse_tables}
\title{Construct a vector of all the available tables}
\usage{
sophisthse_tables(...)
}
\arguments{
\item{...}{further arguments passed into getURL. One may use them to work with proxy.}
}
\value{
vector of all the available tables
}
\description{
Construct a vector of all the available tables
}
\details{
Construct a vector of all the available tables. For the moment contains an error.
BRDATA is a table of tables :) And also some regional data cannot be parsed.
}
\examples{
sophisthse_tables()
}
